Add SQL queries and change DDL so types are numeric

+++ linux_sql/sql/ddl.sql
@@ -6,20 +6,20 @@ CREATE TABLE IF NOT EXISTS host_info (
     num_cpu INTEGER NOT NULL,
     cpu_architecture VARCHAR (255) NOT NULL,
     cpu_model VARCHAR (255) NOT NULL,
-    cpu_mhz REAL NOT NULL,
-    L2_cache INTEGER NOT NULL,
-    total_mem INTEGER NOT NULL,
+    cpu_mhz NUMERIC NOT NULL,
+    L2_cache NUMERIC NOT NULL,
+    total_mem NUMERIC NOT NULL,
     "timestamp" TIMESTAMP NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS host_usage (
     "timestamp" TIMESTAMP NOT NULL,
     host_id INTEGER NOT NULL,
-    memory_free INTEGER NOT NULL,
-    cpu_idle INTEGER NOT NULL,
-    cpu_kernel INTEGER NOT NULL,
-    disk_io INTEGER NOT NULL,
-    disk_available INTEGER NOT NULL,
+    memory_free NUMERIC NOT NULL,
+    cpu_idle NUMERIC NOT NULL,
+    cpu_kernel NUMERIC NOT NULL,
+    disk_io NUMERIC NOT NULL,
+    disk_available NUMERIC NOT NULL,
     FOREIGN KEY (host_id)
         REFERENCES host_info (id)
 );
